Add property config, start decompasition
