+++ ObjectPrinting/PropertyConfig.cs
@@ -1,0 +1,47 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Text;
+
+namespace ObjectPrinting
+{
+    public class PropertyConfig<TOwner, TPropType> : IPropertyPrintingConfig<TOwner, TPropType>
+    {
+        private readonly PrintingConfig<TOwner> printingConfig;
+
+        public PropertyConfig(PrintingConfig<TOwner> printingConfig)
+        {
+            this.printingConfig = printingConfig;
+        }
+
+        public PrintingConfig<TOwner> Using(Func<TPropType, string> print)
+        {
+            return printingConfig;
+        }
+
+        public PrintingConfig<TOwner> Using(CultureInfo culture)
+        {
+            return printingConfig;
+        }
+
+        PrintingConfig<TOwner> IPropertyPrintingConfig<TOwner, TPropType>.ParentConfig => printingConfig;
+    }
+
+    public interface IPropertyPrintingConfig<TOwner, TPropType>
+    {
+        PrintingConfig<TOwner> ParentConfig { get; }
+    }
+    public static class PropertyPrintingConfigExtensions
+    {
+        public static string PrintToString<T>(this T obj, Func<PrintingConfig<T>, PrintingConfig<T>> config)
+        {
+            return config(ObjectPrinter.For<T>()).PrintToString(obj);
+        }
+
+        public static PrintingConfig<TOwner> TrimmedToLength<TOwner>(this PropertyConfig<TOwner, string> propConfig, int maxLen)
+        {
+            return ((IPropertyPrintingConfig<TOwner, string>)propConfig).ParentConfig;
+        }
+
+    }
+}
