+++ ObjectPrinting/Tests/ObjectPrinterAcceptanceTests.cs
@@ -1,4 +1,7 @@
-﻿using NUnit.Framework;
+﻿using System;
+using System.Globalization;
+using System.Threading;
+using NUnit.Framework;
 
 namespace ObjectPrinting.Tests
 {
@@ -10,18 +13,35 @@ namespace ObjectPrinting.Tests
         {
             var person = new Person { Name = "Alex", Age = 19 };
 
-            var printer = ObjectPrinter.For<Person>();
+            var printer = ObjectPrinter.For<Person>()
                 //1. Исключить из сериализации свойства определенного типа
+                .Excluding<Guid>()
                 //2. Указать альтернативный способ сериализации для определенного типа
+                .Printing<int>().Using(i => i.ToString("X"))
                 //3. Для числовых типов указать культуру
+                .Printing<double>().Using(CultureInfo.InvariantCulture)
                 //4. Настроить сериализацию конкретного свойства
                 //5. Настроить обрезание строковых свойств (метод должен быть виден только для строковых свойств)
+                .Printing(p => p.Name).TrimmedToLength(10)
                 //6. Исключить из сериализации конкретного свойства
-            
-            string s1 = printer.PrintToString(person);
+                .Excluding(p => p.Age);
 
-            //7. Синтаксический сахар в виде метода расширения, сериализующего по-умолчанию        
+            var s1 = printer.PrintToString(person);
+
+            //7. Синтаксический сахар в виде метода расширения, сериализующего по-умолчанию
+             
+
             //8. ...с конфигурированием
+            string s3 = person.PrintToString(s => s.Excluding(p => p.Age));
+            Console.WriteLine(s1);
+    
+            Console.WriteLine(s3);
+        }
+
+        [Test]
+        public void PrintPerson_WhenExludeIntFields()
+        {
+            
         }
     }
 }
